testing: separation of concerns
It's been tricky to write tests for the server because there's a
cyclic dependancy between falcore and the filters. for example
It was not possible to test a falcore.Server with a file_filter
because the file_filter lib imported the main falcore lib and thus
the test suite would barf.

In essence the filters have been acting as if they were part of
falcore and not their own individual libraries.

This commit prepares to separate the concerns of falcore into
distinct packages.

 * falcore/buffer
 * falcore/filter (consumes buffer)
 * falcore (consumes buffer,filter)

falcore:

Falcore core library will not change, it's responsibility is network
IO and managing the pipelines and queues of requests, responses
and filters. Falcore will consume falcore/filter.

falcore/filter:

The filter libraries's areas of concern are

 * representing http.Request as a custom data structure
 * manipulating http.Request objects
 * manipulating http.Response object

To keep interoperability the following changes will be made in
subsequent commits:

 * request and response filter packages will be rolled into falcore/filter
 * falcore and falcore/filter will share a common logging interface
 * Request struct and methods moved to falcore/filter
 * RequestFilter/ResponseFilter interfaces moved to falcore/filter
 * Update API/method signatures to reflect the code refactoring