All: refactor to use falcore/filter and falcore/buffer packages
This is the second changeset which refactors falcore to facilitate
easier testing of the server and pipeline. The majority of the
changes are as follows

 * *falcore.Request -> *filter.Request
 * falcore.NewSimpleRequest() -> filter.NewSimpleRequest()
 * etc etc...

No internal method or function code changes other than to update
the calls to create request objects so it's all very familiar.

falcore/filter has a similar logging api as the main falcore package
to maintain compatibility, this needs to have a logger set before
the filter is used. For convencient "func GetLogger() Logger" was
added to falcore so you can simply do as in the example servers and
do filter.SetLogger(fc.GetLogger())

The buffer pool code was updated to have a public API.