tests: refactor filter tests to be stand alone
With the cyclic dependancy on falcore main lib being broken, it
is no longer possible to test the filters using a falcore.Server
because we would have to import falcore taking us back to square
one.

The filter tests in this commit simply use the std library file
and http methods and objects to create appropriate the appropriate
requests and/or responses; and simple call FilterRequest and
FilterResponse on these for evaluation.

In most cases the tests are simpler becuase they don't now require
an http server, this may be of benefit if test break in the future
because there are less 'moving parts'.

A more concrete logger was added to prevent panics in the test
suite from normal error printing.

A server test can now be added to the main library and can import
any filter for testing and evaluation, additionally the server
test quite can have benchmark tests added to it to evaluate the
benefit of code changes that target performance.